main/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-migrate==4.0.4
flask-sqlalchemy==3.0.3
mysqlclient==2.1.1
pika==1.3.1
python-dotenv==0.21.1
requests==2.28.2
sqlalchemy==2.0.4